Implement OptionallyPaginatedEndpoint
